CONTRIB-8596: Rewrite view.php to use a renderer properly
This change also includes changes to the meeting_info web service which
moves the generation of meeting information to a new class
`mod_bigbluebuttonbn\meeting`.

This change needs to be done at the same time so that the template for
the view_page can use the same variabels as when the meeting information
is fetched via the meeting_info and end_meeting web services, where it
is used to update the template.

The effect of this change means a simplification to the javascript for
joining a room.